@@ -1,30 +1,55 @@
-import style from './aboutUs.module.css';
-import AboutUsDog from '../../images/dog1.jpeg';
+import s from './aboutUs.module.css'
+import sprite from '../../images/sprite.svg'
+import AboutUsDog from '../../images/dog_about.png'
 
 export default function AboutUs() {
   return (
-    <div className="Container">
-      <h2 className={style.Title}>О нас</h2>
-      <p className={style.Text}>Компания «Лакмир» начала свою деятельность в 2003 году, а начиная с
-        2008 года, было основано производство лакомств для собак под
-        собственной торговой маркой «Лакмир». На сегодняшний день,
-        мы изготовляем широкий ассортимент лакомств и занимаемся реализацией
-        свежезамороженных кормов для Ваших любимцев.</p>
-      <p className={style.Text}>Миссия компании «Лакмир» - обеспечение потребителей натуральными,
-        качественными и доступными лакомством для домашних любимцев.</p>
-      <div className={style.TextContainer}>
-        <img src={AboutUsDog} alt="dog abouy us page"/>
-        <p className={style.Text}>Лакомства изготовляются из говяжьих субпродуктов, таких как:
-        пенис, ухо, вымя, лёгкое, печенка, нос, трахея, сухожилия, яйца, мясо пищевода;
-        а также из куриных продуктов: крылья, филе, шеи, желудки.
-          <br />
+    <div className={s.container}>
+      <div className={s.wrap}>
+        <h2 className={s.title}>
           <br/>
-        При изготовлении используются только натуральное свежее сырье украинских мясокомбинатов, 
-        которое сопровождается соответствующим свидетельством качества.
-          <br />
           <br/>
-        По окончанию цикла производства, готовая продукция проходит ветеринарный контроль, на основании которого мы получаем соответствующие сертификаты качества
-            и ветеринарное свидетельство на нашу продукцию.</p>
+          <br/>
+          <br/>
+          <br/>
+          <br/>
+          <br/>
+          <br/>
+          <br/>
+          <br/>
+          <br/>
+          <br/>
+          
+          О нас</h2>
+        <div className={s.content_wrap}>
+            <div className={s.about_wrap} data-aos="fade-right">
+              <svg className={s.icon}>
+                  <use href={sprite + "#icon-double_arrow"} alt="arrow" />
+              </svg>
+              <p className={`${s.text} ${s.card}`}>Компания «Лакмир» начала свою деятельность в 2003 году, а начиная с
+                2008 года, было основано производство лакомств для собак под
+                собственной торговой маркой «Лакмир». На сегодняшний день,
+                мы изготовляем широкий ассортимент лакомств и занимаемся реализацией
+                свежезамороженных кормов для Ваших любимцев.
+              <br/>Миссия компании «Лакмир» - обеспечение потребителей натуральными,
+                качественными и доступными лакомством для домашних любимцев.</p>
+            </div>
+            <img src={AboutUsDog} alt="dog about us page" className={s.dog_img}/>
+              {/* <div className={s.textContainer}>
+                
+                <p className={s.text}>Лакомства изготовляются из говяжьих субпродуктов, таких как:
+                пенис, ухо, вымя, лёгкое, печенка, нос, трахея, сухожилия, яйца, мясо пищевода;
+                а также из куриных продуктов: крылья, филе, шеи, желудки.
+                  <br />
+                  <br/>
+                При изготовлении используются только натуральное свежее сырье украинских мясокомбинатов, 
+                которое сопровождается соответствующим свидетельством качества.
+                  <br />
+                  <br/>
+                По окончанию цикла производства, готовая продукция проходит ветеринарный контроль, на основании которого мы получаем соответствующие сертификаты качества
+                    и ветеринарное свидетельство на нашу продукцию.</p>
+              </div> */}
+        </div>
       </div>
     </div>
   );
